+++ packages/react/src/index.tsx
@@ -6,3 +6,5 @@ export * from "./box";
 export * from "./message";
 export * from "./accordion";
 export * from "./dropdown";
+export * from "./modal";
+export * from "./breadcrumb";