+++ packages/react/src/list/link-list.tsx
@@ -3,7 +3,7 @@ import type { ListProps } from "./list";
 import { UnorderedList } from "./list";
 
 export interface LinkListProps extends Omit<ListProps, "listStyle"> {
-  children: React.ReactElement<HTMLLIElement> | React.ReactElement<HTMLLIElement>[];
+  children?: React.ReactElement<HTMLLIElement> | React.ReactElement<HTMLLIElement>[];
 }
 
 export function LinkList(props: LinkListProps) {