+++ packages/react/CHANGELOG.md
@@ -1,5 +1,41 @@
 # @postenbring/hedwig-react
 
+## 0.0.14
+
+### Patch Changes
+
+- 67d4df7: :construction: add WIP modal
+- Updated dependencies [67d4df7]
+  - @postenbring/hedwig-css@0.0.13
+
+## 0.0.13
+
+### Patch Changes
+
+- 6d01b04: Adding component for Bread > Crumbs > Are > Here
+- Updated dependencies [6d01b04]
+  - @postenbring/hedwig-css@0.0.12
+
+## 0.0.12
+
+### Patch Changes
+
+- de791ec: make children prop optional to allow for dangerouslySetInnerHTML usage
+
+## 0.0.11
+
+### Patch Changes
+
+- 2e340d0: :lipstick: update list styling
+- Updated dependencies [2e340d0]
+  - @postenbring/hedwig-css@0.0.11
+
+## 0.0.10
+
+### Patch Changes
+
+- @postenbring/hedwig-css@0.0.10
+
 ## 0.0.9
 
 ### Patch Changes